+++ lib/test/clojure_lsp/feature/code_actions_test.clj
@@ -7,7 +7,9 @@
    [clojure-lsp.test-helper :as h]
    [clojure.java.io :as io]
    [clojure.string :as string]
-   [clojure.test :refer [deftest is testing]]))
+   [clojure.test :refer [deftest is testing]]
+   [matcher-combinators.matchers :as m]
+   [matcher-combinators.test :refer [match?]]))
 
 (h/reset-components-before-test)
 
@@ -34,6 +36,25 @@
                                  []
                                  {:workspace {:workspace-edit true}}
                                  (h/db))))))
+
+(deftest avoid-self-refer-suggestion
+  (h/load-code-and-locs (h/code "(ns baz)"
+                                "(defn func-2 []"
+                                "  (func-1))"
+                                "(defn func-1 [] ())"))
+  (testing "Given a unresolved-var because the function is declared after its use
+            When I ask for code actions
+            Then I don't get an 'Add require' suggestion"
+    (is (match? (m/mismatch [{:title "Add require '[baz :refer [func-1]]'"}])
+                (f.code-actions/all (zloc-of h/default-uri)
+                                    h/default-uri
+                                    3
+                                    4
+                                    [{:code "unresolved-symbol"
+                                      :message "Unresolved symbol: func-1"
+                                      :range {:start {:line 2 :character 3}}}]
+                                    {}
+                                    (h/db))))))
 
 (deftest add-alias-suggestion-code-actions
   (h/load-code-and-locs "(ns clojure.set)" (h/file-uri "file:///clojure.core.clj"))